Factor out api views
